+++ app/src/main/java/com/ztn/app/ui/file/FileActivity.kt
@@ -12,7 +12,7 @@ import com.ztn.app.R
 import com.ztn.app.base.BaseActivity
 import com.ztn.app.base.contract.FileContract
 import com.ztn.app.model.bean.FileBean
-import com.ztn.app.presenter.FilePresent
+import com.ztn.app.presenter.FilePresenter
 import com.ztn.common.ToastHelper
 import kotlinx.android.synthetic.main.activity_files.*
 import java.io.File
@@ -22,7 +22,7 @@ import java.io.File
  * Created by 冒险者ztn on 2019/3/4.
  * 文件列表界面
  */
-class FileActivity : BaseActivity<FilePresent>(), FileContract.View {
+class FileActivity : BaseActivity<FilePresenter>(), FileContract.View {
 
 
     companion object {
